working/requirements.txt
gradio
SpeechRecognition
pyttsx3
google-genai
python-dotenv
pyaudio
